Update for latest App Engine best practices
This change upgrades the code in line with current App Engine best
practices. We upgrade Python to version 2.7, use a WSGI app instead
of the CGI mechanism, switch the templating engine to Jinja2 and
the webapp framework to webapp2. We enable multi-threaded operation,
which ought to lead to improved performance.